8h36

+++ src/main/java/org/example/javablog/services/PostService.java
@@ -12,6 +12,7 @@ import org.example.javablog.model.PostRelationship;
 import org.example.javablog.constant.Visibility;
 import org.example.javablog.repository.PostRelationshipRepository;
 import org.example.javablog.repository.PostRepository;
+import org.example.javablog.repository.RecommendScoreRepository;
 import org.example.javablog.specifications.PostSpecification;
 import org.example.javablog.util.PostUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,6 +42,9 @@ public class PostService {
 
     @Autowired
     private PostRelationshipRepository postRelationshipRepository;
+
+    @Autowired
+    private RecommendScoreRepository recommendScoreRepository;
 
     @Autowired
     private PostUtils postUtils;
@@ -107,14 +111,14 @@ public class PostService {
     public void deletePost(Long postId, Long userId) {
         Post post = blogRepository.findById(postId).orElseThrow(NullPointerException::new);
 
+
         if (!post.getAuthor().getId().equals(userId) && !userService.isAdmin()) {
             throw new SecurityException("User is not authorized to delete this post.");
         }
+        recommendScoreRepository.deleteByPostId(postId);
         blogRepository.delete(post);
     }
     public Page<PostDTO> searchPosts(PostFilterRequest filter, Pageable pageable) throws IOException {
-
-
 
         Long userId = userService.getCurrentUser().getId();
         if (filter.getAuthorId() == null){
@@ -127,7 +131,10 @@ public class PostService {
         }
 
 
-        Specification<Post> spec = PostSpecification.filterBy(filter);
+        Specification<Post> spec = Specification.where(PostSpecification.filterBy(filter))
+                .and(PostSpecification.sortByRecommendScore(userId));
+
+
         Page<PostDTO> posts = postRepository.findAll(spec,pageable).map(PostMapper::toDTO);
 
 
@@ -158,7 +165,7 @@ public class PostService {
 
     public void likePost(Long userId,Long postId){
         if (postUtils.existsByRelationship(userId,postId,PostRelationshipType.LIKED)){
-            return;
+            throw new IllegalArgumentException("You have already liked this post");
         }
         if (postUtils.validateOwnership(userId,postId)){
             throw new IllegalArgumentException("You cannot like your own posts");
@@ -167,6 +174,9 @@ public class PostService {
     }
     @Transactional
     public void unlikePost(Long userId,Long postId){
+        if (!postUtils.existsByRelationship(userId,postId,PostRelationshipType.LIKED)){
+            throw new IllegalArgumentException("You have not liked this post yet");
+        }
         postRelationshipRepository.deleteByUserIdAndPostIdAndPostRelationshipType(userId,postId,PostRelationshipType.LIKED);
     }
     @Transactional
